refactor ``fake_exponential`` to ``_utils`` for general purpose use
