+++ eth/vm/forks/prague/state.py
@@ -2,6 +2,7 @@ from typing import (
     Type,
 )
 
+from eth._utils.calculations import fake_exponential
 from eth.abc import (
     ComputationAPI,
     MessageAPI,
@@ -9,6 +10,7 @@ from eth.abc import (
     TransactionContextAPI,
     TransactionExecutorAPI,
 )
+from eth.exceptions import PyEVMError
 from eth.vm.forks.cancun import (
     CancunState,
 )
@@ -17,7 +19,6 @@ from eth.vm.forks.cancun.constants import (
 )
 from eth.vm.forks.cancun.state import (
     CancunTransactionExecutor,
-    fake_exponential,
 )
 from eth.vm.forks.prague.constants import (
     BLOB_BASE_FEE_UPDATE_FRACTION_PRAGUE,
